performance updates (#1353)
* CurrencyIcon perf updates

* Some pagination updates

* Small re-rendering improvements on left futures sidebar

* Remove unnecessary useMemo in PositionChart

* Restructure components to help re-render perf

* Collapse styles

* Memoize TabButton

* Fix table styling

* Remove unused prop

* Fix debt icon
