+++ components/Button/TabButton.tsx
@@ -16,17 +16,8 @@ export type TabButtonProps = {
 	nofill?: boolean;
 };
 
-const TabButton: React.FC<TabButtonProps> = ({
-	title,
-	detail,
-	badge,
-	active,
-	icon,
-	vertical,
-	nofill,
-	...props
-}) => {
-	return (
+const TabButton: React.FC<TabButtonProps> = React.memo(
+	({ title, detail, badge, active, icon, vertical, nofill, ...props }) => (
 		<StyledButton {...props} {...{ active, vertical, nofill }} noOutline>
 			{!!icon && <div>{icon}</div>}
 			<div>
@@ -38,8 +29,8 @@ const TabButton: React.FC<TabButtonProps> = ({
 				{detail && <p className="detail">{detail}</p>}
 			</div>
 		</StyledButton>
-	);
-};
+	)
+);
 
 const StyledButton = styled(Button)<{
 	active?: boolean;
