+++ sections/futures/MarketInfo/MarketInfo.tsx
@@ -1,55 +1,25 @@
-import Head from 'next/head';
-import { FC } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useRecoilValue } from 'recoil';
+import React from 'react';
 import styled from 'styled-components';
 
-import { DEFAULT_CRYPTO_DECIMALS } from 'constants/defaults';
-import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
-import { currentMarketState, marketAssetRateState } from 'store/futures';
-import { formatCurrency } from 'utils/formatters/number';
-import { getDisplayAsset, isDecimalFour } from 'utils/futures';
+import { MainContent } from 'styles/common';
 
 import MarketDetails from '../MarketDetails';
 import PositionChart from '../PositionChart';
 import UserInfo from '../UserInfo';
+import MarketHead from './MarketHead';
 
-const MarketInfo: FC = () => {
-	const { t } = useTranslation();
+const MarketInfo: React.FC = React.memo(() => (
+	<Container>
+		<MarketHead />
+		<MarketDetails />
+		<PositionChart />
+		<UserInfo />
+	</Container>
+));
 
-	const { selectedPriceCurrency } = useSelectedPriceCurrency();
-
-	const baseCurrencyKey = useRecoilValue(currentMarketState);
-	const basePriceRate = useRecoilValue(marketAssetRateState);
-
-	const marketName = getDisplayAsset(baseCurrencyKey);
-
-	return (
-		<Container>
-			<Head>
-				<title>
-					{basePriceRate
-						? t('futures.market.page-title-rate', {
-								marketName,
-								rate: formatCurrency(selectedPriceCurrency.name, basePriceRate, {
-									currencyKey: selectedPriceCurrency.name,
-									minDecimals:
-										marketName != null && isDecimalFour(marketName)
-											? DEFAULT_CRYPTO_DECIMALS
-											: undefined,
-								}),
-						  })
-						: t('futures.market.page-title')}
-				</title>
-			</Head>
-			<MarketDetails />
-			<PositionChart />
-			<UserInfo />
-		</Container>
-	);
-};
-
-const Container = styled.div`
+const Container = styled(MainContent)`
+	margin: unset;
+	max-width: unset;
 	padding-bottom: 20px;
 `;
